Reduced code by 227 Lines: Extract all transformation‐related functions (matrix creation, object mirroring, and applying transforms) into a dedicated module. This not only removes clutter from your main file but also ensures a single source of truth for coordinate conversion.
